possibly breaking code with MemoryTile update
