+++ src/TMGE/Tile.java
@@ -2,24 +2,11 @@ package TMGE;
 
 import java.awt.Color;
 
-public class Tile {
+abstract class Tile {
 	Color color;
-	boolean active = false;
 	
 	public Tile(Color c){
 		color = c;
-	}
-	
-	public boolean isActive() {
-		return active;
-	}
-	
-	public void reveal() {
-		active = true;
-	}
-	
-	public void hide() {
-		active = false;
 	}
 	
 	public Color getColor() {
